+++ src/main/java/kitchenpos/application/OrderService.java
@@ -8,10 +8,10 @@ import kitchenpos.domain.OrderTable;
 import kitchenpos.repository.MenuRepository;
 import kitchenpos.repository.OrderRepository;
 import kitchenpos.repository.OrderTableRepository;
-import kitchenpos.ui.dto.request.OrderChangeStatusRequest;
-import kitchenpos.ui.dto.request.OrderCreateRequest;
-import kitchenpos.ui.dto.request.OrderLineItemCreateRequest;
-import kitchenpos.ui.dto.response.OrderResponse;
+import kitchenpos.application.dto.request.OrderChangeStatusRequest;
+import kitchenpos.application.dto.request.OrderCreateRequest;
+import kitchenpos.application.dto.request.OrderLineItemCreateRequest;
+import kitchenpos.application.dto.response.OrderResponse;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.util.CollectionUtils;
