+++ src/main/java/kitchenpos/application/dto/response/OrderTableInTableGroupResponse.java
@@ -1,4 +1,4 @@
-package kitchenpos.ui.dto.response;
+package kitchenpos.application.dto.response;
 
 import kitchenpos.domain.OrderTable;
