+++ src/main/java/kitchenpos/ui/OrderRestController.java
@@ -1,9 +1,9 @@
 package kitchenpos.ui;
 
 import kitchenpos.application.OrderService;
-import kitchenpos.ui.dto.request.OrderChangeStatusRequest;
-import kitchenpos.ui.dto.request.OrderCreateRequest;
-import kitchenpos.ui.dto.response.OrderResponse;
+import kitchenpos.application.dto.request.OrderChangeStatusRequest;
+import kitchenpos.application.dto.request.OrderCreateRequest;
+import kitchenpos.application.dto.response.OrderResponse;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
