+++ src/test/java/kitchenpos/api/order/OrderChangeOrderStatusApiTest.java
@@ -2,9 +2,9 @@ package kitchenpos.api.order;
 
 import kitchenpos.api.config.ApiTestConfig;
 import kitchenpos.domain.OrderStatus;
-import kitchenpos.ui.dto.request.OrderChangeStatusRequest;
-import kitchenpos.ui.dto.response.OrderLineItemResponse;
-import kitchenpos.ui.dto.response.OrderResponse;
+import kitchenpos.application.dto.request.OrderChangeStatusRequest;
+import kitchenpos.application.dto.response.OrderLineItemResponse;
+import kitchenpos.application.dto.response.OrderResponse;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
