+++ src/test/java/kitchenpos/api/table/TableChangeNumberOfGuestsApiTest.java
@@ -1,8 +1,8 @@
 package kitchenpos.api.table;
 
 import kitchenpos.api.config.ApiTestConfig;
-import kitchenpos.ui.dto.request.OrderTableChangeGuestNumberRequest;
-import kitchenpos.ui.dto.response.OrderTableResponse;
+import kitchenpos.application.dto.request.OrderTableChangeGuestNumberRequest;
+import kitchenpos.application.dto.response.OrderTableResponse;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
